feat: get budget, calc total balance
